refactor(core): move colour enums into Core/Enums for docs discovery
