+++ docs/content/docs/advanced/enums.mdx
@@ -43,9 +43,10 @@ layout primitives.
 
 ## Text & sizing
 
-`Lattice\Lattice\Ui\Enums` — sizing and color for `Text` and `Icon`.
+Sizing lives in `Lattice\Lattice\Ui\Enums`; the color vocabulary lives in
+`Lattice\Lattice\Core\Enums` and backs the `Color` value object.
 
-<EnumTable enums={["Size", "Color"]} />
+<EnumTable enums={["Size", "ColorName", "ColorKind"]} showNamespace />
 
 The icon names a `Lattice\Lattice\Ui\Enums\Icon` accepts are the enum's own cases — one per bundled
 SVG. The [Icons](/core/icons/) page covers where they come from and how to add your own.
