+++ src/Core/Enums/ColorKind.php
@@ -1,7 +1,7 @@
 <?php
 declare(strict_types=1);
 
-namespace Lattice\Lattice\Core;
+namespace Lattice\Lattice\Core\Enums;
 
 use Lattice\Lattice\Attributes\TypeScript;
